+++ include/jfif.h
@@ -1,26 +1,41 @@
-//=============================================================
+// ------------------------------------------------------------------------------
 //
-// Copyright (c) 2010-2026 Simon Southwell
-// All rights reserved.
+//  File Name:           jfif.h
+//  Design Unit Name:    Header for JFIF API
+//  Revision:            OSVVM MODELS STANDARD VERSION
 //
-// Date: 18th January 2010
+//  Maintainer:          Simon Southwell      email:  simon.southwell@gmail.com
+//  Contributor(s):
+//     Simon Southwell   simon.southwell@gmail.com
 //
-// This file is part of JFIF.
+//  Description:
+//      Header for JFIF API
 //
-// JFIF is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
+//  Developed by:
+//        Simon Southwell
 //
-// JFIF is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
+//  Revision History:
+//    Date      Version    Description
+//    06/2026   2026.08   Initial revision
 //
-// You should have received a copy of the GNU General Public License
-// along with JFIF. If not, see <http://www.gnu.org/licenses/>.
+//  This file is part of OSVVM.
 //
-//=============================================================
+//  Copyright (c) 2026 by Simon Southwell
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+// ------------------------------------------------------------------------------
+//
 //
 // The code implements decoding of JFIF/JPEG data based on the
 // following standards (links supplied)
@@ -34,7 +49,7 @@
 // This file provides the definitions for external code to link
 // to the main JFIF/JPEG decode routine jpeg_process_jfif().
 //
-//=============================================================
+// ------------------------------------------------------------------------------
 
 #include <stdint.h>