+++ include/jfif_gtk.h
@@ -16,7 +16,7 @@
 //
 //  Revision History:
 //    Date      Version    Description
-//    06/2026   2026.07   Initial revision
+//    06/2026   2026.08   Initial revision
 //
 //  This file is part of OSVVM.
 //