+++ include/rv32_cpu_gdb.h
@@ -15,7 +15,7 @@
 //
 //  Revision History:
 //    Date      Version    Description
-//    09/2025   ????.??    Update model to v1.3.6
+//    07/2026   2026.08    Initial revision
 //    01/2023   2023.01    Released with OSVVM CoSim
 //    5th Aug   2021      Earlier version
 //