+++ tests/rv32demo/rv32_cosim_utils.h
@@ -20,7 +20,7 @@
 //
 //  Revision History:
 //    Date      Version    Description
-//    07/2026   ????.??    Initial revision
+//    07/2026   2026.08    Initial revision
 //
 //
 //  This file is part of OSVVM.